XWIKI-13368 * correct @since version * move layout handler interface to internal package
