+++ xwiki-platform-core/xwiki-platform-rendering/xwiki-platform-rendering-macros/xwiki-platform-rendering-macro-code/src/main/java/org/xwiki/rendering/macro/code/CodeMacroParameters.java
@@ -59,6 +59,7 @@ public class CodeMacroParameters extends BoxMacroParameters
 
     /**
      * @param layout the layout to apply to the code.
+     * @since 11.4RC1
      */
     @PropertyDescription("the layout format (plain or with line numbers)")
     public void setLayout(CodeMacroLayout layout)
@@ -68,6 +69,7 @@ public class CodeMacroParameters extends BoxMacroParameters
 
     /**
      * @return the layout to apply to the code.
+     * @since 11.4RC1
      */
     public CodeMacroLayout getLayout()
     {
